% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notion.R
\name{convert_notionapp_link}
\alias{convert_notionapp_link}
\title{Convert notion url to app link}
\usage{
convert_notionapp_link(url)
}
\arguments{
\item{url}{url (character)}
}
\description{
Convert notion url to app link
}
\examples{
\dontrun{
convert_notionapp_link("https://www.notion.so/728460a1ba514e6c9a3cfde8a8c371a5")
}
}
