% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{detect_tweet_latest_since_id}
\alias{detect_tweet_latest_since_id}
\title{Get the latest since_id of tweets for a query in a file}
\usage{
detect_tweet_latest_since_id(dir, query)
}
\arguments{
\item{dir}{A character string indicating the directory path where tweet data is stored.}

\item{query}{A character string indicating the query used to search for tweets.}
}
\value{
A tibble with two columns, "since_id" and "file".
"since_id" column contains the latest since_id for the query,
and "file" column contains the name of the file that contains the tweet data.
}
\description{
Get the latest since_id of tweets for a query in a file
}
\examples{
\dontrun{
detect_tweet_latest_since_id("data-raw", "rstats")
}
}
