% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbd.R
\name{get_openbd}
\alias{get_openbd}
\title{Request to openBD API}
\usage{
get_openbd(isbn)
}
\arguments{
\item{isbn}{isbn code}
}
\description{
Request to openBD API
}
\examples{
\dontrun{
get_openbd("9784478108536")
}
}
