% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanji_ruby.R
\name{html_yomi}
\alias{html_yomi}
\title{html showing the reading of the word}
\usage{
html_yomi(x, yomi, style = "space")
}
\arguments{
\item{x}{word}

\item{yomi}{character}

\item{style}{style for yomi text}
}
\description{
html showing the reading of the word
}
\examples{
html_yomi("CRAN", "The Comprehensive R Archive Network", "parenthesis")
}
