% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{line_notify}
\alias{line_notify}
\title{Line Notification}
\usage{
line_notify(token, message, img_file = NULL, .disabled = FALSE)
}
\arguments{
\item{token}{Personal access token}

\item{message}{text}

\item{img_file}{path to upload image file}

\item{.disabled}{option}
}
\description{
Line Notification
}
