% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_ds.R
\name{scale_color_ds}
\alias{scale_color_ds}
\alias{scale_fill_ds}
\title{Color scale for data science plot}
\usage{
scale_color_ds(
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  na.value = "gray60",
  ...
)

scale_fill_ds(
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  na.value = "gray60",
  ...
)
}
\arguments{
\item{palette}{main}

\item{discrete}{discrete variables}

\item{reverse}{color order}

\item{na.value}{na color}

\item{...}{path to other functions}
}
\description{
Color scale for data science plot
}
