% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_df.R
\name{summary_df}
\alias{summary_df}
\title{Tidy summary for data.frame}
\usage{
summary_df(data, .trim = NULL, na.rm = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{.trim}{Option.}

\item{na.rm}{a logical value indicating whether NA values should be
stripped before the computation proceeds.}
}
\description{
Alternative broom::tidy() for data.frame.
}
