% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{theme_ds_dark}
\alias{theme_ds_dark}
\title{Theme of data science for dark mode}
\usage{
theme_ds_dark(base_size = 11, ...)
}
\arguments{
\item{base_size}{base font size}

\item{...}{Other arguments passed to [theme_light][ggplot2::theme_light]}
}
\description{
Theme of data science for dark mode
}
