% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{write_search_tweets_rds}
\alias{write_search_tweets_rds}
\title{Search recent tweets and save to RDS}
\usage{
write_search_tweets_rds(dir, query, since_id, ...)
}
\arguments{
\item{dir}{A character string indicating the directory path where tweet data is stored.}

\item{query}{A character string indicating the query used to search for tweets.}

\item{since_id}{A character string. This argument is used to retrieve only
tweets that have been posted after a specific tweet.
If `since_id` is `NA`, it is set to `NULL`.}

\item{...}{Additional arguments to be passed to the `write_search_tweets_rds` function.}
}
\description{
Search recent tweets and save to RDS
}
\examples{
\dontrun{
write_search_tweets_rds("data-raw", "rstats", since_id = "1621390633146667008")
}
}
