% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{write_twitter_query_status}
\alias{write_twitter_query_status}
\title{Write twitter query status to CSV}
\usage{
write_twitter_query_status(dir, query)
}
\arguments{
\item{dir}{A character string indicating the directory path where tweet data is stored.}

\item{query}{A character string indicating the query used to search for tweets.}
}
\description{
This function retrieves the latest `since_id` for each query and writes the results to a CSV file.
}
\examples{
\dontrun{
write_twitter_query_status("data-raw", c("#rstats", "#datascience"))
}
}
