% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanji_ruby.R
\name{kanji_ruby}
\alias{kanji_ruby}
\title{Ruby in html}
\usage{
kanji_ruby(x, view = FALSE)
}
\arguments{
\item{x}{word}

\item{view}{View html output in the viewer}
}
\description{
Ruby in html
}
\examples{
kanji_ruby(intToUtf8(c(35576L, 34892L, 28961L, 24120L)))
}
